update fitur dan fix bug penelitian

+++ app/Http/Controllers/HewanController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Models\Hewan;
 use App\Models\Pelanggan;
+use App\Models\Penitipan;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
+use Illuminate\Validation\ValidationException;
 
 class HewanController extends Controller
 {
@@ -14,9 +16,23 @@ class HewanController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $data = Hewan::latest()->orderBy('nama_hewan', 'asc')->paginate(5);
+        $keyword = $request->input('keyword');
+
+        $data = Hewan::where(function ($query) use ($keyword) {
+            $query->where('nama_hewan', 'LIKE', '%' . $keyword . '%')
+                ->orWhere('jenis_hewan', 'LIKE', '%' . $keyword . '%')
+                ->orWhere('umur', 'LIKE', '%' . $keyword . '%')
+                ->orWhere('berat', 'LIKE', '%' . $keyword . '%')
+                ->orWhereHas('pelanggan', function ($query) use ($keyword) {
+                    $query->where('nama_pelanggan', 'LIKE', '%' . $keyword . '%');
+                });
+        })
+            ->with('pelanggan') // Menyertakan relasi dengan model Pelanggan
+            ->latest()
+            ->orderBy('nama_hewan', 'asc')
+            ->paginate(5);
         return view('hewan.index')->with('data', $data);
     }
 
@@ -146,6 +162,12 @@ class HewanController extends Controller
      */
     public function destroy($id)
     {
+
+        $penitipan = Penitipan::where('hewan_id', $id)->first();
+        if ($penitipan) {
+            //jika ada data hewan yang sedang dititipkan, maka hentikan proses delete
+            throw ValidationException::withMessages(['hewan_id' => 'Data Hewan Sedang dititipkan di penitipan']);
+        }
         Hewan::where('id', $id)->delete();
         return redirect('/hewan')->with('success', 'Berhasil Hapus Data!!');
     }
